organizando correo

+++ db/migrate/20181026052233_add_price_to_property.rb
@@ -1,5 +1,5 @@
 class AddPriceToProperty < ActiveRecord::Migration[5.2]
   def change
-    add_column :properties, :price, :long
+    add_column :properties, :price, :float
   end
 end
